Requisito 4 v1.1
