+++ src/pages/Home.js
@@ -1,15 +1,31 @@
 import React, { Component } from 'react';
 import Category from '../components/Category';
+import * as api from '../services/api';
 
 class Home extends Component {
+  constructor() {
+    super();
+    this.state = {
+      categories: [],
+    };
+  }
+
+  componentDidMount() { this.fetchCategories(); }
+
+  async fetchCategories() {
+    const categoriesArr = await api.getCategories().then((categories) => categories);
+    this.setState({ categories: categoriesArr });
+  }
+
   render() {
+    const { categories } = this.state;
     return (
       <div>
         <input type="search" />
         <p data-testid="home-initial-message">
           Digite algum termo de pesquisa ou escolha uma categoria.
         </p>
-        <Category />
+        <Category categories={ categories } />
       </div>
     );
   }
